separate out user and global plugins

+++ cli/hooks/plugins.js
@@ -9,14 +9,11 @@ module.exports = async({config}) => {
   const {hyperdrive} = config;
   // get some config we need
   const {plugin} = hyperdrive.config.get();
-
-  // if we dont have a global plugin manifest then create it
-  if (!fs.existsSync(plugin.globalManifest) && has(hyperdrive, 'lando.pluginDirs')) {
-    debug('no global plugin manifest at %o looking for plugins in %o', plugin.globalManifest, hyperdrive.lando.pluginDirs);
-    // look for plugins
+  // helper to return plugin information
+  const getPlugins = (dirs = [], type = 'global') => {
     const Plugin = hyperdrive.getClass('plugin');
-    const globalPlugins = get(hyperdrive, 'lando.pluginDirs', [])
-    .filter(dir => dir.type === 'global')
+    return dirs
+    .filter(dir => dir.type === type)
     .map(dir => ({type: dir.type, dirs: hyperdrive.bootstrap.findPlugins(dir.dir, dir.depth)}))
     .map(dirs => dirs.dirs.map(dir => new Plugin({root: dir, type: dirs.type})))
     .flat(Number.POSITIVE_INFINITY)
@@ -27,11 +24,23 @@ module.exports = async({config}) => {
       const {channel, config, debug, installDir, pjson, root, updateAvailable, ...dumpable} = plugin;
       return dumpable;
     });
+  };
 
+  // if we dont have a global plugin manifest then create it
+  if (!fs.existsSync(plugin.globalManifest) && has(hyperdrive, 'lando.pluginDirs')) {
+    debug('no global plugin manifest at %o looking for plugins in %o', plugin.globalManifest, hyperdrive.lando.pluginDirs);
+    // look for global plugins
+    const globalPlugins = getPlugins(get(hyperdrive, 'lando.pluginDirs', []), 'global');
     // do the dump
     fs.mkdirSync(path.dirname(plugin.globalManifest), {recursive: true});
     fs.writeFileSync(plugin.globalManifest, JSON.stringify(globalPlugins, null, 2));
   }
+
+  // get our user plugins
+  // @NOTE: user plugins are manually added so it is necessary to scan for them ever time
+  const userPlugins = getPlugins(get(hyperdrive, 'lando.pluginDirs', []), 'user');
+  debug('discovered %o user %o plugins', userPlugins.length, get(hyperdrive, 'lando.product', 'lando'));
+  hyperdrive.plugins.add('user', {type: 'literal', store: hyperdrive.bootstrap.normalizePlugins(userPlugins)});
 
   // get our global plugins
   // @TODO: try/catch for require?
